Move `dashboard` screens to a seperate group folder for clarity
