+++ app/(app)/(dashboard)/user.tsx
@@ -17,11 +17,11 @@ import {
   Spacer,
 } from "tamagui";
 import { useEffect, useState } from "react";
-import LessonCard from "../../src/components/LessonCard";
-import { ExercisePopover } from "../../src/components/ExercisePopover";
-import { ExerciseTypes } from "../../src/utils/enums";
-import { Exercise } from "../../src/models/Exercise";
-import { ExerciseService } from "../../src/services/ExerciseService";
+import LessonCard from "../../../src/components/LessonCard";
+import { ExercisePopover } from "../../../src/components/ExercisePopover";
+import { ExerciseTypes } from "../../../src/utils/enums";
+import { Exercise } from "../../../src/models/Exercise";
+import { ExerciseService } from "../../../src/services/ExerciseService";
 
 export default function Dashboard() {
   const DEBUG_USER_UUID = "ebabaa6c-4254-465e-9f2f-f285a2364277";
